+++ packages/picasso-provider/src/Picasso/PicassoGlobalStylesProvider.tsx
@@ -1,11 +1,13 @@
 import type { ReactNode, ForwardRefExoticComponent, RefAttributes } from 'react'
-import React, { useRef, useState, useCallback } from 'react'
+import React, { useEffect, useRef, useState, useCallback } from 'react'
 
 import type { RootContextProps } from './RootContext'
 import { RootContext } from './RootContext'
 import type { EnvironmentType, TextLabelProps } from '../types'
 import type { PicassoRootNodeProps } from './PicassoRootNode'
 import { isBrowser } from '../utils'
+import type { BreakpointKeys } from './config/breakpoints'
+import { useScreens } from './config/breakpoints'
 
 export interface PicassoGlobalStylesProviderProps extends TextLabelProps {
   children?: ReactNode
@@ -14,6 +16,14 @@ export interface PicassoGlobalStylesProviderProps extends TextLabelProps {
   >
   environment: EnvironmentType<'test' | 'temploy'>
   disableTransitions?: boolean
+}
+
+const breakpointKeyByRange: Record<BreakpointKeys, BreakpointKeys> = {
+  xs: 'xs',
+  sm: 'sm',
+  md: 'md',
+  lg: 'lg',
+  xl: 'xl',
 }
 
 const PicassoGlobalStylesProvider = (
@@ -29,8 +39,13 @@ const PicassoGlobalStylesProvider = (
 
   const [picassoRootMounted, setPicassoRootMounted] = useState(false)
   const rootRef = useRef<HTMLDivElement | null>(null)
+
+  const screens = useScreens<BreakpointKeys>()
+  const currentBreakpointRange = screens(breakpointKeyByRange)
+
   const [contextValue, setContextValue] = useState<RootContextProps>({
     rootRef,
+    currentBreakpointRange,
     hasTopBar: false,
     setHasTopBar: (hasTopBar: boolean) => {
       setContextValue({
@@ -56,6 +71,15 @@ const PicassoGlobalStylesProvider = (
     },
     disableTransitions,
   })
+
+  useEffect(() => {
+    if (contextValue.currentBreakpointRange !== currentBreakpointRange) {
+      setContextValue({
+        ...contextValue,
+        currentBreakpointRange,
+      })
+    }
+  }, [currentBreakpointRange])
 
   const setRootRef = useCallback(
     (ref: HTMLDivElement) => {