+++ packages/picasso/src/Radio/test.tsx
@@ -103,7 +103,7 @@ describe('Radio', () => {
 
     it('renders radio in a grid group', () => {
       const { container } = render(
-        <Radio.Group horizontal small={4} name='my-group'>
+        <Radio.Group horizontal sm={4} name='my-group'>
           <Radio label='LABEL+1' value='VALUE+1' />
           <Radio label='LABEL+2' value='VALUE+2' />
           <Radio label='LABEL+3' value='VALUE+3' />