+++ packages/picasso/src/RadioGroup/RadioGroup.tsx
@@ -1,28 +1,22 @@
 /* eslint-disable react/no-array-index-key */
 import React from 'react'
-import type { GridSize } from '@material-ui/core/Grid'
 import type { RadioGroupProps } from '@material-ui/core'
 import { RadioGroup as MUIRadioGroup } from '@material-ui/core'
 import type { Theme } from '@material-ui/core/styles'
 import { makeStyles, useTheme } from '@material-ui/core/styles'
 
+import type { GridSizeProps } from '../GridItem'
 import type { GridProps } from '../Grid'
 import { GridCompound as Grid } from '../GridCompound'
 import styles from './styles'
 
 type GridSpacing = GridProps['spacing']
 
-export interface Props extends RadioGroupProps {
+export interface Props extends RadioGroupProps, GridSizeProps {
   /** Align radios horizontally  */
   horizontal?: boolean
   /** Defines amount of space between radio buttons components (in px) */
   spacing?: GridSpacing
-  /** Defines the number of grids the component is going to use. It's applied for all the screen sizes with the lowest priority */
-  small?: boolean | GridSize
-  /** Defines the number of grids the component is going to use. It's applied for the medium breakpoint and wider screens */
-  medium?: boolean | GridSize
-  /** Defines the number of grids the component is going to use. It's applied for the large breakpoint and wider screens */
-  large?: boolean | GridSize
 }
 
 // Using { index: -1 } to inject CSS link to the bottom of the head
@@ -34,7 +28,7 @@ const useStyles = makeStyles<Theme>(styles, {
 })
 
 const RadioGroup = (props: Props) => {
-  const { horizontal, spacing, small, medium, large, ...rest } = props
+  const { horizontal, spacing, xs, sm, md, lg, xl, ...rest } = props
   const { grid: gridClass, gridItem: gridItemClass, ...classes } = useStyles()
   const { spacing: themeSpacing } = useTheme()
 
@@ -54,9 +48,11 @@ const RadioGroup = (props: Props) => {
           <Grid.Item
             key={index}
             className={gridItemClass}
-            small={small}
-            medium={medium}
-            large={large}
+            xs={xs}
+            sm={sm}
+            md={md}
+            lg={lg}
+            xl={xl}
           >
             {child}
           </Grid.Item>