+++ packages/picasso/src/Slider/story/Controlled.example.tsx
@@ -12,7 +12,7 @@ const Example = () => {
 
   return (
     <Grid alignItems='center'>
-      <Grid.Item small={6}>
+      <Grid.Item sm={6}>
         <Grid alignItems='center'>
           <Grid.Item>
             <Button
@@ -24,7 +24,7 @@ const Example = () => {
               Zoom out
             </Button>
           </Grid.Item>
-          <Grid.Item small>
+          <Grid.Item sm>
             <Slider value={value} onChange={handleChange} max={5} />
           </Grid.Item>
           <Grid.Item>