+++ packages/picasso-pictograms/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-shared@12.0.0
+
 ## 0.6.0
 
 ### Minor Changes